CSS the living thing out of this

+++ server/formserver/bibweb/bibweb.js
@@ -1,16 +1,16 @@
-/*
-module.exports = function (app) {
-  require("./student")(app);
-  require("./author")(app);
-  require("./book")(app);
-  require("./lending")(app);
-};
-*/
-
 const connection = require("../db");
+
+const allowedTableNames = ["student", "author", "lending", "book"];
+
+const isNotAllowedTableName = (value) => !allowedTableNames.includes(value);
 
 module.exports = function (app) {
   app.get("/bibweb/:table", (req, res) => {
+    if (isNotAllowedTableName(req.params.table)) {
+      res.status(400).send({ message: "Bad table name" });
+      return;
+    }
+
     connection.query(
       `SELECT * FROM bibweb.${req.params.table};`,
       (err, result) => {
@@ -24,6 +24,11 @@ module.exports = function (app) {
   });
 
   app.post("/bibweb/:table", (req, res) => {
+    if (isNotAllowedTableName(req.params.table)) {
+      res.status(400).send({ message: "Bad table name" });
+      return;
+    }
+
     connection.query(
       `INSERT INTO bibweb.${req.params.table} SET ?`,
       { ...req.body },
@@ -38,6 +43,11 @@ module.exports = function (app) {
   });
 
   app.put("/bibweb/:table", (req, res) => {
+    if (isNotAllowedTableName(req.params.table)) {
+      res.status(400).send({ message: "Bad table name" });
+      return;
+    }
+
     connection.query(
       `UPDATE bibweb.${req.params.table} SET ? WHERE id=?`,
       [{ ...req.body }, req.body.id],
@@ -52,6 +62,11 @@ module.exports = function (app) {
   });
 
   app.delete("/bibweb/:table/:id", (req, res) => {
+    if (isNotAllowedTableName(req.params.table)) {
+      res.status(400).send({ message: "Bad table name" });
+      return;
+    }
+
     connection.query(
       `DELETE FROM bibweb.${req.params.table} WHERE id=?;`,
       req.params.id,
